Renamed convertBlobAttributes.ts => blobToBinary.ts
